+++ Modular Gameplay Overhaul/Modules/Professions/Extensions/TreeExtensions.cs
@@ -4,6 +4,7 @@
 
 using DaLion.Shared.Extensions;
 using Microsoft.Xna.Framework;
+using NetFabric.Hyperlinq;
 using StardewValley.TerrainFeatures;
 
 #endregion using directives