+++ Modular Gameplay Overhaul/Modules/Core/CHANGELOG.md
@@ -1,5 +1,11 @@
 ﻿# Modular Overhaul Core Change Log
 
+## 1.3.1
+
+### Fixed
+
+* Improvements to Chinese localization.
+
 ## 1.3.0
 
 ### Added